+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/dcBotModule.kt
@@ -3,8 +3,14 @@ package io.github.sophon.discord
 import dev.kord.core.Kord
 import io.github.sophon.adminModule
 import io.github.sophon.core.coreModule
+import io.github.sophon.core.feature.Config
+import io.github.sophon.data.ReportRepo
+import io.github.sophon.discord.data.FileManager
 import io.github.sophon.discord.data.InMemoryGlossaryDB
+import io.github.sophon.discord.data.JsonReportRepo
 import io.github.sophon.discord.domain.DiscordButtonBuilder
+import io.github.sophon.discord.domain.Tracker
+import io.github.sophon.discord.domain.TrackerImpl
 import io.github.sophon.discord.featureRegistry.featureRegistryModule
 import io.github.sophon.discord.usecase.CreateEmbedUseCase
 import io.github.sophon.discord.usecase.CreateErrorEmbedBuilderUseCase
@@ -13,12 +19,14 @@ import io.github.sophon.discord.usecase.CreateMutableEmbedUseCase
 import io.github.sophon.discord.usecase.CreatePlainMessageUseCase
 import io.github.sophon.discord.usecase.CreateReplyEmbedUseCase
 import io.github.sophon.discord.usecase.HandleButtonInteractionUseCase
+import io.github.sophon.discord.usecase.PostDailyReportEmbedUseCase
 import io.github.sophon.discord.usecase.ResultToEmbedUseCase
 import io.github.sophon.discord.usecase.RouteCommandToFeatureUseCase
 import io.github.sophon.dreamcancel.dreamCancelModule
 import io.github.sophon.ewgfModule
 import io.github.sophon.glossaryinfil.data.GlossaryDB
 import io.github.sophon.glossaryinfil.infilModule
+import io.github.sophon.statsModule
 import io.github.sophon.wikiSuperCombo.superComboModule
 import io.github.sophon.wikidustloop.dustLoopModule
 import io.github.sophon.wikimizuumi.mizuumiModule
@@ -43,6 +51,7 @@ fun initKoin(
         coreModule,
         dcBotModule(kord),
         adminModule(),
+        statsModule(),
 
         infilModule,
         wavuModule(),
@@ -65,6 +74,16 @@ fun dcBotModule(kord: Kord) = module {
 
     singleOf(::DiscordBotImpl).bind<DiscordBot>()
 
+    single {
+        TrackerImpl(
+            statsFeatureInfo = get(),
+            statsChannelId = get<Config>().statsConfig?.statsChannelIdList?.firstOrNull() ?: "",
+            scheduler = get(),
+            scope = get(),
+            statsTracker = get(),
+        )
+    }.bind<Tracker>()
+
     singleOf(::RouteCommandToFeatureUseCase)
     singleOf(::CreateErrorEmbedBuilderUseCase)
     singleOf(::CreatePlainMessageUseCase)
@@ -74,8 +93,12 @@ fun dcBotModule(kord: Kord) = module {
     singleOf(::ResultToEmbedUseCase)
     singleOf(::CreateMutableEmbedUseCase)
     singleOf(::HandleButtonInteractionUseCase)
+    singleOf(::PostDailyReportEmbedUseCase)
 
     singleOf(::DiscordButtonBuilder)
 
     singleOf(::InMemoryGlossaryDB).bind<GlossaryDB>()
+
+    singleOf(::FileManager)
+    singleOf(::JsonReportRepo).bind<ReportRepo>()
 }