+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/domain/Scheduler.kt
@@ -9,10 +9,13 @@ import kotlin.time.Duration.Companion.hours
 
 class Scheduler {
     fun <T>start(
+        initialDelay: Duration = Duration.ZERO,
         period: Duration = TIME_UPDATE_INTERVAL_H.hours,
         task: suspend () -> T,
     ): Flow<T> {
         return flow {
+            delay(initialDelay)
+
             while (true) {
                 emit(task.invoke())
                 delay(period)