+++ feat/stats/src/commonMain/kotlin/io/github/sophon/domain/usecase/RecordUseCase.kt
@@ -1,0 +1,18 @@
+package io.github.sophon.domain.usecase
+
+import io.github.sophon.core.domain.EmptyResult
+import io.github.sophon.core.domain.ExcludeFromCoverage
+import io.github.sophon.core.domain.Result
+import io.github.sophon.domain.StatsError
+import io.github.sophon.domain.model.Command
+
+@ExcludeFromCoverage("TODO: cover with test")
+internal class RecordUseCase() {
+    fun invoke(
+        command: Command,
+        cache: MutableMap<Command, Long>,
+    ): EmptyResult<StatsError> {
+        cache[command] = cache.getOrElse(key = command, defaultValue = { 0L }) + 1
+        return Result.Success(Unit)
+    }
+}